pyproject.toml
[project]
name = "arkterm"
version = "0.1.0-dev.2"
description = "Shell Shocked: Wire an LLM Directly into Your Linux Terminal"
authors = [{ name = "Saadman Rafat", email = "saadman@outlook.com" }]
readme = "README.md"
requires-python = ">=3.10"
keywords = [
    "ai",
    "cli",
    "terminal",
    "llm",
    "assistant",
    "shell",
    "groq",
    "chat"
]


classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries",
  "Topic :: Utilities"
]
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.3.5",
]

[tool.hatch.build.targets.wheel]
packages = ["arkterm"]

[project.scripts]
arkterm = "arkterm.main:main"